refactor: extract StorageHelper.java
